# Conflicts:
#	Client/Pages/MovieDetails.razor
#	Client/Services/DbAccess.cs

+++ Client/Models/Movies/Movie.cs
@@ -54,7 +54,7 @@ namespace Client.Models
             public string? ReleaseDate;
 
             [JsonProperty("revenue")]
-            public int? Revenue;
+            public Int64? Revenue;
 
             [JsonProperty("runtime")]
             public int? Runtime;
